Update `demo` example station A with naive clean function.
